rename configparser to rtparse - to avoid import-issues with the python-configparser module
